% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestDiffusion.R
\name{ForestDiffusion.generate}
\alias{ForestDiffusion.generate}
\title{Generate new observations with a trained ForestDiffusion model}
\usage{
ForestDiffusion.generate(object, batch_size = NULL, n_t = NULL, seed = NULL)
}
\arguments{
\item{object}{a ForestDiffusion object}

\item{batch_size}{(optional) number of observations generated; if not provided, will generate as many observations as the original dataset}

\item{n_t}{(optional) number of noise levels (and sampling steps); increase for higher performance, but slows down training and sampling; if not provided, will use the same n_t as used in training.}

\item{seed}{(optional) random seed used}
}
\value{
Returns a data.frame with the generated data
}
\description{
Generate new observations by solving the reverse SDE (vp) / ODE (flow) starting from pure Gaussian noise.
}
\examples{
 \dontrun{
 data(iris)
 X = data.frame(iris[,1:4])
 y = iris[,5]
 
 ## Generation
 
 Xy = X
 Xy$y = y
 forest_model = ForestDiffusion(X=Xy, n_cores=1, n_t=50, duplicate_K=50, flow=TRUE)
 Xy_fake = ForestDiffusion.generate(forest_model, batch_size=NROW(Xy))
}  

 # Simple test (to verify that everything works)
 data(iris)
 set.seed(1)
 X = data.frame(iris[1:5,1:2])
 X = missForest::prodNA(X, noNA = 0.5)
 forest_model = ForestDiffusion(X=X, n_cores=1, n_t=2, duplicate_K=1, flow=TRUE)
 Xy_fake = ForestDiffusion.generate(forest_model, batch_size=NROW(X))

}
\references{
Alexia Jolicoeur-Martineau, Kilian Fatras, Tal Kachman. 
Generating and Imputing Tabular Data via Diffusion and Flow-based Gradient-Boosted Trees. 
arXiv:2309.09968.
}
