% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestDiffusion.R
\name{with_datasets}
\alias{with_datasets}
\title{Evaluate an expression in multiple generated/imputed datasets}
\usage{
with_datasets(data, expr)
}
\arguments{
\item{data}{a list of datasets}

\item{expr}{An expression to evaluate for each imputed data set. Formula's
containing a dot (notation for "all other variables") do not work.}
}
\value{
An object of S3 class mira
}
\description{
It performs a computation for each dataset (function modified from mice::with.mids). For example, you can use this function to train a different glm model per dataset and then pool the estimates (akin to with multiple imputations, but more general so that it can be applied to any dataset).
}
\examples{
\dontrun{
library(mice)

# Load iris
data(iris)
Xy = data.frame(iris[,1:4])
Xy$y = iris[,5]

# add missing data
Xy = missForest::prodNA(Xy, noNA = 0.2)

forest_model = ForestDiffusion(X=Xy, n_cores=1, n_t=50, duplicate_K=50, flow=FALSE)
nimp = 5 # number of imputations needed
# regular (fast)
Xy_imp = ForestDiffusion.impute(forest_model, k=nimp)
# REPAINT (slow, but better)
Xy_imp = ForestDiffusion.impute(forest_model, repaint=TRUE, r=10, j=5, k=nimp)

# Fit a model per imputed dataset
fits <- with_datasets(Xy_imp, glm(y ~ Sepal.Length, family = 'binomial'))

# Pool the results
mice::pool(fits) 
}

# Simple test (to verify that everything works)
data(iris)
set.seed(1)
X = data.frame(iris[1:5,1:2])
X = missForest::prodNA(X, noNA = 0.5)
forest_model = ForestDiffusion(X=X, n_cores=1, n_t=2, duplicate_K=1, flow=FALSE)
Xy_imp = ForestDiffusion.impute(forest_model, repaint=TRUE, r=2, j=5, k=1)
Xy_imp = ForestDiffusion.impute(forest_model, k=5)
fits <- with_datasets(Xy_imp, lm(Sepal.Width ~ Sepal.Length))
mice::pool(fits) 

}
