% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestDiffusion.R
\name{ForestDiffusion.generate}
\alias{ForestDiffusion.generate}
\title{Generate new observations with a trained ForestDiffusion model}
\usage{
ForestDiffusion.generate(object, batch_size = NULL, n_t = NULL)
}
\arguments{
\item{object}{a ForestDiffusion object}

\item{batch_size}{(optional) number of observations generated; if not provided, will generate as many observations as the original dataset}

\item{n_t}{(optional) number of noise levels (and sampling steps); increase for higher performance, but slows down training and sampling; if not provided, will use the same n_t as used in training.}
}
\value{
Returns a data.frame with the generated data
}
\description{
Generate new observations by solving the reverse SDE (vp) / ODE (flow) starting from pure Gaussian noise.
}
\examples{
 \dontrun{
 data(iris)
 X = data.frame(iris[,1:4])
 y = iris[,5]
 
 ## Generation
 
 Xy = X
 Xy$y = y
 forest_model = ForestDiffusion(Xy, n_t=50, duplicate_K=50, flow=TRUE)
 Xy_fake = ForestDiffusion.generate(forest_model, batch_size=NROW(Xy))
}  

}
\references{
Alexia Jolicoeur-Martineau, Kilian Fatras, Tal Kachman. 
Generating and Imputing Tabular Data via Diffusion and Flow-based Gradient-Boosted Trees. 
arXiv:2309.09968.
}
